pyproject.toml
[project]
name = "document-ai"
version = "0.1.0"
authors = [
    { name = "Angelos Panagiotopoulos", email = "angelospanag@protonmail.com" }
]
description = "A simple FastAPI application that allows users to upload pdf or docx documents, get a summary generated by a local LLM via Ollama, and ask questions about their content."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.7",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = ["pytest>=8.4.1", "ruff>=0.12.10"]

[tool.ruff]
lint.select = [
    "E", "W",    # pycodestyle
    "F",         # pyflakes (undefined names, etc.)
    "B",         # bugbear
    "I",         # isort
    "N",         # pep8-naming
    "SIM",       # simplify
    "UP",        # pyupgrade
    "PERF",      # perflint
    "RET",       # flake8-return
    "RSE",       # flake8-raise
]
